net-snmp: update to 5.8

+++ release/src/router/net-snmp-5.7.2/snmplib/snmp_alarm.c
@@ -319,7 +319,7 @@ set_an_alarm(void)
         signal(SIGALRM, alarm_handler);
         setitimer(ITIMER_REAL, &it, NULL);
         DEBUGMSGTL(("snmp_alarm", "schedule alarm %d in %ld.%03ld seconds\n",
-                    nextalarm, (long) delta.tv_sec, (delta.tv_usec / 1000)));
+                    nextalarm, (long) delta.tv_sec, (long)(delta.tv_usec / 1000)));
 # else  /* HAVE_SETITIMER */
 #  ifdef SIGALRM
         signal(SIGALRM, alarm_handler);
@@ -446,7 +446,7 @@ snmp_alarm_register_hr(struct timeval t, unsigned int flags,
 
     DEBUGMSGTL(("snmp_alarm",
                 "registered alarm %d, t = %ld.%03ld, flags=0x%02x\n",
-                (*s)->clientreg, (long) (*s)->t.tv_sec, ((*s)->t.tv_usec / 1000),
+                (*s)->clientreg, (long) (*s)->t.tv_sec, (long)((*s)->t.tv_usec / 1000),
                 (*s)->flags));
 
     if (start_alarms) {
